+++ src/app/common/business-models/property.business-model.ts
@@ -8,4 +8,5 @@ export interface PropertyBusinessModel {
   type: string;
   address: AddressBusinessModel;
   leaseContractId: string;
+  fixtureInventoryId: string;
 }