+++ src/app/features/fixture-inventory/services/post-fixture-inventory.model.ts
@@ -1,0 +1,7 @@
+export interface PostFixtureInventoryModel {
+  arrivalFixtureInventoryDate: Date,
+  exitFixtureInventoryDate: Date,
+  arrivalComments: string,
+  exitComments: string,
+  propertyId: string;
+}