+++ CHANGELOG.md
@@ -14,6 +14,20 @@ Nomenclature :
 
 ### Unreleased
 
+#### Added
+
+### [0.1.4] - 2023-02-29
+
+#### Added
+
+- Updated db.json with fixture-inventory data to complete the simulated api
+- Added GET /fixtures-inventory/:id and display it in a card
+- Added property-card to removed duplicated when we reminded property in fixture-detail
+- Added POST /fixtures-inventory (form)
+- Added PUT /fixtures-inventory/:id (form)
+- Added DELETE /fixtures-inventory/:id
+- Added fixture-inventory and lease-contrat columns on properties-dashboard
+
 ### [0.1.3] - 2023-02-27
 
 #### Added